:heavy_check_mark: removes child state and adds remove function
